Demo project
Tasks (operations) flow control plays an important role in many software projects. Its reliable and effective implementation may be vital for project success. Several years ago, I already presented an article with a similar scope entitled Tiny Framework for Parallel Computing. Although the main approach of that work remained unchanged, some concepts are evolving, and design is remarkably different. Implementation of major types was simplified, and additional important and useful features were added to the infrastructure. My intention was to make this article self-sustained and independent on its predecessor, so some concepts are reiterated. The subject of this framework is control of processes consisting of separate operations performing either sequentially or in parallel. Examples of such processes can be machine control, complex software testing, financial flows, etc. Code is written in C# and targeted to .NET Core, but can be easily ported to other languages. It was tested in Windows and Linux (Ubuntu 18.04).